Move helpers to their own dir to speedup loading
Knife loads anything in the lib/chef/knife/ dir so if we want to keep
something from loading each time we can stick it in a subdir.
