+++ lib/chef/knife/helpers/azure_base.rb
