+++ lib/chef/knife/helpers/azurerm_base.rb
